+++ lib/lightning_web/live/project_live/settings.ex
@@ -62,10 +62,26 @@ defmodule LightningWeb.ProjectLive.Settings do
         project_user.user_id == current_user.id
       end)
 
-    can_create_webhook_auth_method =
+    can_write_webhook_auth_method =
       Permissions.can?(
         ProjectUsers,
-        :create_webhook_auth_method,
+        :write_webhook_auth_method,
+        current_user,
+        project_user
+      )
+
+    can_write_github_connection =
+      Permissions.can?(
+        ProjectUsers,
+        :write_github_connection,
+        current_user,
+        project_user
+      )
+
+    can_initiate_github_sync =
+      Permissions.can?(
+        ProjectUsers,
+        :initiate_github_sync,
         current_user,
         project_user
       )
@@ -74,14 +90,6 @@ defmodule LightningWeb.ProjectLive.Settings do
       Permissions.can?(
         ProjectUsers,
         :create_project_credential,
-        current_user,
-        project_user
-      )
-
-    can_edit_webhook_auth_method =
-      Permissions.can?(
-        ProjectUsers,
-        :edit_webhook_auth_method,
         current_user,
         project_user
       )
@@ -106,9 +114,8 @@ defmodule LightningWeb.ProjectLive.Settings do
        can_edit_project_name: can_edit_project_name,
        can_edit_project_description: can_edit_project_description,
        can_edit_data_retention: project_user.role in [:owner, :admin],
-       can_create_webhook_auth_method: can_create_webhook_auth_method,
+       can_write_webhook_auth_method: can_write_webhook_auth_method,
        can_create_project_credential: can_create_project_credential,
-       can_edit_webhook_auth_method: can_edit_webhook_auth_method,
        show_github_setup: show_github_setup,
        show_repo_setup: show_repo_setup,
        show_sync_button: show_sync_button,
@@ -117,16 +124,10 @@ defmodule LightningWeb.ProjectLive.Settings do
        branches: [],
        loading_branches: false,
        github_enabled: VersionControl.github_enabled?(),
-       can_install_github: can_install_github(socket),
+       can_install_github: can_write_github_connection,
+       can_initiate_github_sync: can_initiate_github_sync,
        selected_credential_type: nil
      )}
-  end
-
-  defp can_install_github(socket) do
-    case socket.assigns.project_user.role do
-      :viewer -> false
-      _ -> true
-    end
   end
 
   defp repo_settings(%Project{id: project_id}) do
@@ -402,14 +403,22 @@ defmodule LightningWeb.ProjectLive.Settings do
      )}
   end
 
-  def handle_event("run_sync", params, %{assigns: %{current_user: u}} = socket) do
-    case VersionControl.run_sync(params["id"], u.email) do
-      {:ok, :fired} ->
-        {:noreply, socket |> put_flash(:info, "Sync Initialized")}
+  def handle_event(
+        "initiate_sync",
+        params,
+        %{assigns: %{current_user: u}} = socket
+      ) do
+    if socket.assigns.can_initiate_github_sync do
+      case VersionControl.initiate_sync(params["id"], u.email) do
+        {:ok, :fired} ->
+          {:noreply, socket |> put_flash(:info, "Sync Initialized")}
 
-      _err ->
-        # we should log or instrument this situation
-        {:noreply, socket |> put_flash(:error, "Sync Error")}
+        _err ->
+          # we should log or instrument this situation
+          {:noreply, socket |> put_flash(:error, "Sync Error")}
+      end
+    else
+      {:noreply, socket |> put_flash(:error, "Viewers Cannot Initiate Sync")}
     end
   end