+++ test/lightning_web/live/project_live_test.exs
@@ -2,6 +2,7 @@ defmodule LightningWeb.ProjectLiveTest do
   use LightningWeb.ConnCase, async: false
 
   alias Lightning.Repo
+  alias Lightning.Name
 
   import Phoenix.LiveViewTest
   import Lightning.ProjectsFixtures
@@ -1313,22 +1314,12 @@ defmodule LightningWeb.ProjectLiveTest do
       end)
     end
 
-    test "all project users can see the project webhook auth methods", %{
-      conn: conn
-    } do
+    test "all project users can see the project webhook auth methods" do
       project = insert(:project)
       auth_methods = insert_list(4, :webhook_auth_method, project: project)
 
-      for project_user <-
-            Enum.map([:editor, :admin, :owner, :viewer], fn role ->
-              insert(:project_user,
-                role: role,
-                project: project,
-                user: build(:user)
-              )
-            end) do
-        conn = log_in_user(conn, project_user.user)
-
+      for conn <-
+            build_project_user_conns(project, [:editor, :admin, :owner, :viewer]) do
         {:ok, _view, html} =
           live(
             conn,
@@ -1341,7 +1332,7 @@ defmodule LightningWeb.ProjectLiveTest do
       end
     end
 
-    test "authorized project users can add a new project webhook auth method",
+    test "owners/admins can add a new project webhook auth method, editors/viewers can't",
          %{
            conn: conn
          } do
@@ -1350,16 +1341,7 @@ defmodule LightningWeb.ProjectLiveTest do
       settings_path =
         Routes.project_project_settings_path(conn, :index, project.id)
 
-      for project_user <-
-            Enum.map([:editor, :admin, :owner], fn role ->
-              insert(:project_user,
-                role: role,
-                project: project,
-                user: build(:user)
-              )
-            end) do
-        conn = log_in_user(conn, project_user.user)
-
+      for conn <- build_project_user_conns(project, [:owner, :admin]) do
         {:ok, view, _html} =
           live(
             conn,
@@ -1386,7 +1368,7 @@ defmodule LightningWeb.ProjectLiveTest do
                |> element("form#choose_auth_type_form_#{modal_id}")
                |> has_element?()
 
-        credential_name = "#{project_user.role}credentialname"
+        credential_name = Name.generate()
 
         refute render(view) =~ credential_name
 
@@ -1415,6 +1397,22 @@ defmodule LightningWeb.ProjectLiveTest do
           )
 
         assert html =~ credential_name
+      end
+
+      for conn <- build_project_user_conns(project, [:editor, :viewer]) do
+        {:ok, view, _html} =
+          live(
+            conn,
+            settings_path
+          )
+
+        assert view
+               |> element("button#add_new_auth_method:disabled")
+               |> has_element?()
+
+        modal_id = "new_auth_method_modal"
+
+        refute view |> element("##{modal_id}") |> has_element?()
       end
     end
 
@@ -1445,7 +1443,7 @@ defmodule LightningWeb.ProjectLiveTest do
       refute view |> element("#new_auth_method_modal") |> has_element?()
     end
 
-    test "authorized project users can add edit a project webhook auth method",
+    test "owners/admins can add edit a project webhook auth method, editors/viewers can't",
          %{
            conn: conn
          } do
@@ -1460,16 +1458,7 @@ defmodule LightningWeb.ProjectLiveTest do
       settings_path =
         Routes.project_project_settings_path(conn, :index, project.id)
 
-      for project_user <-
-            Enum.map([:editor, :admin, :owner], fn role ->
-              insert(:project_user,
-                role: role,
-                project: project,
-                user: build(:user)
-              )
-            end) do
-        conn = log_in_user(conn, project_user.user)
-
+      for conn <- build_project_user_conns(project, [:owner, :admin]) do
         {:ok, view, _html} =
           live(
             conn,
@@ -1484,7 +1473,7 @@ defmodule LightningWeb.ProjectLiveTest do
 
         assert view |> element("##{modal_id}") |> has_element?()
 
-        credential_name = "#{project_user.role}credentialname"
+        credential_name = Name.generate()
 
         refute render(view) =~ credential_name
 
@@ -1509,6 +1498,24 @@ defmodule LightningWeb.ProjectLiveTest do
           )
 
         assert html =~ credential_name
+      end
+
+      for conn <- build_project_user_conns(project, [:editor, :viewer]) do
+        {:ok, view, _html} =
+          live(
+            conn,
+            settings_path
+          )
+
+        assert view
+               |> element(
+                 "a#edit_auth_method_link_#{auth_method.id}.cursor-not-allowed"
+               )
+               |> has_element?()
+
+        modal_id = "edit_auth_#{auth_method.id}_modal"
+
+        refute view |> element("##{modal_id}") |> has_element?()
       end
     end
 
@@ -1539,7 +1546,9 @@ defmodule LightningWeb.ProjectLiveTest do
         )
 
       assert view
-             |> element("a#edit_auth_method_link_#{auth_method.id}")
+             |> element(
+               "a#edit_auth_method_link_#{auth_method.id}.cursor-not-allowed"
+             )
              |> has_element?()
 
       refute view
@@ -1562,7 +1571,7 @@ defmodule LightningWeb.ProjectLiveTest do
 
       project_user =
         insert(:project_user,
-          role: :editor,
+          role: :admin,
           project: project,
           user: build(:user)
         )
@@ -1639,7 +1648,7 @@ defmodule LightningWeb.ProjectLiveTest do
 
       project_user =
         insert(:project_user,
-          role: :editor,
+          role: :admin,
           project: project,
           user: build(:user)
         )
@@ -1702,11 +1711,11 @@ defmodule LightningWeb.ProjectLiveTest do
     end
   end
 
-  test "authorized project users can delete a project webhook auth method",
+  test "owners and admins can delete a project webhook auth method",
        %{conn: conn} do
     project = insert(:project)
 
-    for role <- [:owner, :admin, :editor] do
+    for role <- [:owner, :admin] do
       auth_method =
         insert(:webhook_auth_method,
           project: project,
@@ -1737,6 +1746,10 @@ defmodule LightningWeb.ProjectLiveTest do
 
       modal_id = "delete_auth_#{auth_method.id}_modal"
 
+      assert view
+             |> element("#delete_auth_method_#{modal_id}_#{auth_method.id}")
+             |> has_element?()
+
       view
       |> form("#delete_auth_method_#{modal_id}_#{auth_method.id}",
         delete_confirmation_changeset: %{confirmation: "DELETE"}
@@ -1751,6 +1764,42 @@ defmodule LightningWeb.ProjectLiveTest do
 
       assert flash["info"] ==
                "Your Webhook Authentication method has been deleted."
+    end
+
+    for role <- [:editor, :viewer] do
+      auth_method =
+        insert(:webhook_auth_method,
+          project: project,
+          auth_type: :basic
+        )
+
+      project_user =
+        insert(:project_user,
+          role: role,
+          project: project,
+          user: build(:user)
+        )
+
+      settings_path =
+        Routes.project_project_settings_path(conn, :index, project.id)
+
+      conn = log_in_user(conn, project_user.user)
+
+      {:ok, view, _html} =
+        live(
+          conn,
+          settings_path
+        )
+
+      refute view
+             |> element("a#delete_auth_method_link_#{auth_method.id}")
+             |> has_element?()
+
+      modal_id = "delete_auth_#{auth_method.id}_modal"
+
+      refute view
+             |> element("#delete_auth_method_#{modal_id}_#{auth_method.id}")
+             |> has_element?()
     end
   end